+++ sktime/dists_kernels/algebra.py
@@ -5,7 +5,7 @@ __author__ = ["fkiraly"]
 import numpy as np
 
 from sktime.base import _HeterogenousMetaEstimator
-from sktime.dists_kernels._base import BasePairwiseTransformerPanel
+from sktime.dists_kernels.base import BasePairwiseTransformerPanel
 
 SUPPORTED_MTYPES = ["pd-multiindex", "nested_univ", "df-list", "numpy3D"]
