+++ sktime/dists_kernels/base/__init__.py
@@ -1,0 +1,8 @@
+"""Module exports for dist_kernels.base module."""
+
+from sktime.dists_kernels.base._base import (
+    BasePairwiseTransformer,
+    BasePairwiseTransformerPanel,
+)
+
+__all__ = ["BasePairwiseTransformer", "BasePairwiseTransformerPanel"]
